// component/homepage/type.ts
export type FeatureType = {
  id: number;
  title: string;
  description: string;
  image: string;
};

export type PresentationType = {
  title: string;
  subTitle: string;
  icon_0: string;
  title1: string;
  icon_1: string;
  title2: string;
  icon_2: string;
  title3: string;
  icon_3: string;

}

export type SiteFeatureType = {
  id: number;
  title: string;
  description: string;
  logo: string;
  backImage: string;
  url: string;
}